+++ src/main/java/com/reboot_course/firstcome_system/order/domain/entity/Order.java
@@ -1,0 +1,45 @@
+package com.reboot_course.firstcome_system.order.domain.entity;
+
+import com.reboot_course.firstcome_system.common.entity.BaseEntity;
+import com.reboot_course.firstcome_system.member.entity.Member;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+import java.math.BigDecimal;
+
+@Entity
+@Getter
+@SuperBuilder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name = "`order`")
+public class Order extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "order_id")
+    private Integer id;
+
+    @Column(nullable = false, precision = 10, scale = 2)
+    private BigDecimal totalAmount;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private OrderStatus status;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id", nullable = false, foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
+    private Member member;
+
+    public void cancel() {
+        this.status = OrderStatus.CANCELLED;
+    }
+
+    public void returned() {
+        this.status = OrderStatus.RETURNED;
+    }
+}
