--- src/main/java/com/reboot_course/firstcome_system/order/dto/internal/OrderPaging.java
@@ -1,10 +1,0 @@
-package com.reboot_course.firstcome_system.order.dto.internal;
-
-import lombok.Builder;
-
-import java.util.List;
-
-
-@Builder
-public record OrderPaging(List<Integer> ids, String nextCursor) {
-}
