--- src/main/java/com/reboot_course/firstcome_system/order/dto/request/create/OrderCreateItem.java
@@ -1,16 +1,0 @@
-package com.reboot_course.firstcome_system.order.dto.request.create;
-
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotNull;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class OrderCreateItem {
-    @NotNull
-    private Integer productId;
-
-    @Min(1)
-    private Integer quantity;
-}
