--- src/main/java/com/reboot_course/firstcome_system/order/dto/response/read/OrderMainResponse.java
@@ -1,14 +1,0 @@
-package com.reboot_course.firstcome_system.order.dto.response.read;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-@Builder
-public class OrderMainResponse {
-    private List<OrderMainItem> items;
-    private String nextCursor;
-    private long totalCount;
-}
