+++ src/main/java/com/reboot_course/firstcome_system/order/order/dto/internal/OrderPaging.java
@@ -1,0 +1,10 @@
+package com.reboot_course.firstcome_system.order.order.dto.internal;
+
+import lombok.Builder;
+
+import java.util.List;
+
+
+@Builder
+public record OrderPaging(List<Integer> ids, String nextCursor) {
+}
