+++ src/main/java/com/reboot_course/firstcome_system/order/order/dto/request/create/OrderCreateItem.java
@@ -1,0 +1,16 @@
+package com.reboot_course.firstcome_system.order.order.dto.request.create;
+
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class OrderCreateItem {
+    @NotNull
+    private Integer productId;
+
+    @Min(1)
+    private Integer quantity;
+}
