+++ src/main/java/com/reboot_course/firstcome_system/product/usecase/ProductReader.java
@@ -1,8 +1,8 @@
 package com.reboot_course.firstcome_system.product.usecase;
 
 import com.reboot_course.firstcome_system.common.utils.CursorUtils;
-import com.reboot_course.firstcome_system.order.dto.request.create.OrderCreateItem;
-import com.reboot_course.firstcome_system.order.vo.OrderProductMap;
+import com.reboot_course.firstcome_system.order.order.dto.request.create.OrderCreateItem;
+import com.reboot_course.firstcome_system.order.order.vo.OrderProductMap;
 import com.reboot_course.firstcome_system.product.dto.response.ProductItemDTO;
 import com.reboot_course.firstcome_system.product.dto.response.ProductMainResponse;
 import com.reboot_course.firstcome_system.product.entity.Product;
